fix `.line` pragma ignoring column information
Summary
=======

Use the column information provided by the tuple value, which fixes,
for example, stack traces produced by assertion failures in compile-time
contexts pointing to the wrong column.

In addition, the tuple passed to the `.line` pragma is now verified to
have the correct number of elements, which fixes the compiler crashing
when it's a tuple with only a single element.

Details
=======

In `pragmaLine`:
- don't allow `nkPar` -- all literal tuple construction expressions have
  to use `nkTupleConstr` nodes
- require the provided tuple node to have exactly three elements
- use more descriptive names instead of `x` and `y`
- take the column information from the third tuple element
- build the line information via `newLineInfo`, which also makes sure
  that out-of-range line and column numbers are properly handled (by
  clamping them)
- use the correct node when creating errors and use `wrapError` for
  proper propagation

In addition, the error message for when the provided tuple value doesn't
have the correct shape is slightly improved, now detailing how the
tuple is required to look like.

A test (`t9768.nim`) that relied on incorrect column being set is
adjusted.