internal: don't instantiate concepts when lifting types
Summary
=======

Don't instantiate invocations of generic concepts when lifting types
(`liftParamType`). Instead, only turn the invocation into a
`tyUserTypeClassInst` type.

This removes a usage of `semtypinst`'s "meta-types allowed" mode, and is
thus a step towards removing the latter.

Details
=======

Instead of using `instGenericContainer(allowedMetaTypes = true)` for
lifting the invocation of generic concepts into a `tyUserTypeClassInst`,
the invocation is now copied, transitioned to a `tyUserTypeClassInst`,
and then has the underlying `tyUserTypeClass` the invocation applies to
added as the last element.

The above is **not** the same as what `instGenericContainer` did. Most
importantly, generic invocations used as arguments to the invocation
are not instantiated. In practice, this means that a parameter type like
`Concept[Generic[T]]` will reach `typeRel` and subsequently
`matchUserTypeClass` as

 `(UserTypeClassInst (...) (GenericInvocation (...) (...)) (...))`

instead of as

`(UserTypeClassInst (...) (GenericInst (...) (...) (...)) (...))`

The `tyGenericInvocation` has the `tfHasMeta` flag included and a
`tyInferred` is thus later created when matching against the concept,
which is correct. Since `typeRel` does support resolving `tyInferred`
types that have a `tyGenericInvocation` as the base, this doesn't cause
problems.