make `.cursor` fields work according to spec (#863)
## Summary

Don't inject lifetime hook calls for `.cursor` fields when lifting
hooks, making the `.cursor` pragma on object fields behave as described
in the specification.

This also fixes double frees and other ownership issues for closure
iterators that contained `.cursor` locals.

## Details

The specification says that the `.cursor` annotation prevents
construction/destruction pairs, and while this was already true for all
cursor locals and assignments to cursor fields, inside lifted hooks it
was only true for `ref` and `proc` fields when `--gc:arc|orc` was used.

The special-casing of `ref` and `proc` and in `fillBodyObj` is removed,
meaning that hook calls are now always prevented for all cursor fields.

A direct test for cursor fields is added plus a test for cursor locals
lifted into closure iterator environments (where the previous behaviour
caused double frees for non-`ref`/non-`proc` fields).