warn if `untyped` params followed by non-`untyped` (#862)
## Summary

Emit a warning if an `untyped`, or `varargs[untyped]`, parameter is
followed by a non-`untyped` parameter. This will eventually become an
error, as it violates the principle where previously analysed AST cannot
be "unanalysed" during overload dispatch resolution.

## Details

`semtypes.semProcTypeNode` to emit a warning if an `untyped`, or
`varargs[untyped]`, parameter is followed by a non-`untyped` parameter.
This is a new warning message so new entries were required in legacy
reports, unfortunately.

The manual has been updated to indicate the future direction where this
will become an error.
Finally added a test to ensure that warnings are generated as required.

As part of converting it to an error the warning will be used to track
down all APIs/signatures that require updating.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>