perform constant folding prior to `sempass2`
All places where AST is either passed on to `sempass2` or
`transf` now run the constant-folding pass first. This means that
`sempass2` can now rely on constant-folded AST, which allows for
removing the tentative folding in call analysis.

In addition, `semcomptime.checkAux` doesn't have to ignore
`mLengthArray` calls anymore, as those are now folded away already.

Finally, the inlining of constants in `transf` is removed and the
module's documentation adjusted. Removing obsolete `nkCheckedFieldExpr`
node is done in the folding pass, meaning that `transf` doesn't have
to do it anymore.