store a `tyGenericInvocation` in `tyCompositeTypeClass`
Instead of storing a fully instantiated meta-type (`tyGenericInst`
containing type variables) in the last position of `tyCompositeTypeClass`,
use a `tyGenericInvocation`. This removes the last usage of `semtypinst`'s
"meta-types allowed" mode.

Rendering of `tyCompositeTypeClass` (`typesrenderer`), their type-
relation logic (`sigmatch`), and `inferWithMetatype` are adjusted to
account for this change.

Not skipping the first generic instance of a `tyCompositeTypeClass` in
`typeRel` also fixes a type relation bug with generic types.