make the "log all VM bytecode" debug feature work (#868)
## Summary

Correctly treat specifying no value for the `expandVmListing`
conditional symbol as meaning "log all VM bytecode", making the debug
feature work as described in the documentation.

## Details

Testing for the "no value specified" case was missing in `logBytecode`.
Specifying no value for a conditional symbol result in it being assigned
the string "true", so detection has to take this into account.

This does mean that `--define:expandVmListing` and
`--define:expandVmListing:true` mean the same thing, but given that
'true' is seldomly used as a procedure name, this is unlikely to be a
problem in practice.