restore the `privateAccess` feature
The traversal implemented in `toObjectFromRefPtrGeneric` was wrong,
but that so far didn't cause any issues, since instantiated object
types always used the symbol of the generic type.

With instantiated object types now having a unique symbol, invoking
`privateAccess` with an instantiated type only allows access to that
singular instance's field.