sem: correctly compute size for partial objects
Summary
=======

Treat partial objects as having an unknown size until they're completed.
This prevents their usage in compile-time contexts prior to being
completed and fixes memory corruption issues with the C target, which
were caused by the size computed for the partial object types being
wrong.

Since the JS and VM targets don't support `sizeof`/`alignof`/`offsetof`
for incomplete types, attempting to call either magic for an incomplete
partial object will result in compiler errors.

Details
=======

The size and alignment for all types is computed and set for all types
at the end of a type section, and this led to the size of the incomplete
package-level (i.e., partial) objects being set already.

When a package-level object is started (that is, first added to the
`packageType` list), its size and alignment is now set to 'unknown',
preventing the value from being incorrectly computed and disabling
folding of `sizeof`/`alignof`/`offsetof` magics. Once finished (i.e.,
the final record AST is analyzed), the size and alignments values are
set to 'uncomputed'.

Unfortunately, this means that other aggregate types embedding partial
objects prior to them being completed are also treated as incomplete,
and, since there's currently no way to propagate the completion of a
partial object to all its users, stay as incomplete, preventing their
use with `sizeof`/`alignof`/`offset` at compile-time and with the JS and
VM targets.