nimsuggest: remove `v1` implementation (#881)
## Summary

Remove the implementation of the legacy v1 `nimsuggest` version. This
simplifies the code and thus helps with restructuring/refactoring
`nimsuggest`, which is in the process of moving to the language-server
protocol (LSP).

The `--v1` CLI option is removed from `nimsuggest`, making the `--v2`
option, which is kept for backwards compatibility, the only remaining
version option.

## Details

Since there's only a single version left, the fields for tracking the
selected version are obsolete. `suggestVersion` having the value 0 meant
that the `v2` version is selected, while a value of 1 meant that
the now-removed `v1` version is selected.

* remove the `suggestVersion` field and all its usages
* remove the `version` field from the `Suggest` type
* remove the `findUsages` procedure, which was only used with version
  `v1`