stdlib: temporarily add AST pattern matching (#882)
## Summary

Add the `ast_pattern_matching` module to the experimental standard
library. It implements pattern matching for `NimNode` AST and is copied
from https://github.com/krux02/ast-pattern-matching.

The first iteration of the language-server protocol (LSP) implementation
is going to use the pattern matching, but it's likely that the module
will be removed from the standard library again.

## Details

A copy of the `ast-pattern-matching` implementation already existed
within the `tests` directory, where it is used by the `tastpec.nim`
test. The module is moved to the `lib/experimental` directory and the
import is adjusted.